+++ src/view/position-box/boxes/AutoBoxDebug.svelte
@@ -2,10 +2,11 @@
    import {
       applyPosition,
       applyStyles,
-      draggable,
-      resizeObserver }  from '@typhonjs-fvtt/runtime/svelte/action';
+      resizeObserver }        from '@typhonjs-fvtt/runtime/svelte/action';
 
-   import { boxStore }  from '../boxStore.js';
+   import { draggableGsap }   from '@typhonjs-fvtt/runtime/svelte/gsap';
+
+   import { boxStore }        from '../boxStore.js';
 
    export let box;
 
@@ -40,7 +41,7 @@
 
 <div class=box
      use:applyPosition={box.position}
-     use:draggable={{ position: box.position }}
+     use:draggableGsap={{ position: box.position, inertia: true }}
      use:resizeObserver={box.position}
      style:background={box.color}>
    {#if $labels}{@html cornersText}{/if}